postfork/flask-gunicorn/requirements.txt
click==8.1.7
Flask==2.3.3
googleapis-common-protos==1.52.0
grpcio==1.56.2
gunicorn==22.0.0
itsdangerous==2.1.2
Jinja2==3.1.4
MarkupSafe==2.1.3
# opentelemetry-api==1.20.0
# opentelemetry-exporter-otlp==1.20.0
# opentelemetry-instrumentation==0.41b0
# opentelemetry-instrumentation-flask==0.41b0
# opentelemetry-instrumentation-wsgi==0.41b0
# opentelemetry-sdk==1.20.0
protobuf==3.20.3
six==1.15.0
thrift==0.13.0
uWSGI==2.0.22
Werkzeug==3.0.6
wrapt==1.16.0
middleware-io